requirements.txt
openai==1.93.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.175.0
dotenv==0.9.9
elevenlabs==2.6.0
Pillow==11.3.0
pydub==0.25.1
moviepy==2.2.1